+++ app/assets/javascripts/ovto_app/actions.js.rb
@@ -9,6 +9,27 @@ class MyApp < Ovto::App
     include ProjectActions
     include TaskActions
 
+    def fetch(*args)
+      started_at = Time.now
+      promise = Promise.new
+      actions.increment_connection(started_at: started_at)
+      Ovto.fetch(*args).then{|*args|
+        actions.decrement_connection(started_at: started_at)
+        promise.resolve(*args)
+      }.fail{|e|
+        promise.reject(e)
+      }
+      return promise
+    end
+
+    def increment_connection(state:, started_at:)
+      return {n_connections: state.n_connections + 1}
+    end
+
+    def decrement_connection(state:, started_at:)
+      return {n_connections: state.n_connections - 1}
+    end
+
     def show_upcoming_tasks(state:)
       return {main_view: MainViewInfo.new(type: :upcoming_tasks)}
     end