+++ app/assets/javascripts/ovto_app/actions/project_actions.js.rb
@@ -4,7 +4,7 @@ class MyApp < Ovto::App
   class Actions < Ovto::Actions
     module ProjectActions
       def get_projects(state:)
-        Ovto.fetch('/projects.json').then {|json|
+        fetch('/projects.json').then {|json|
           actions.receive_projects(projects: json.map{|x| Project.from_json(x)})
         }.fail {|e|
           console.log("get_projects", e)